Move static assets to public
