+++ public/api/kubernetes/apis/README.md
@@ -1,0 +1,4 @@
+# Static mock files, representing k8s API server responces
+
+- URL paths are prefixed by `/api/kubernetes`
+- GET, POST, PATCH, PUT share the same resources path
