+++ src/components/Utils/ConditionLabel/__mocks__/conditionsMock.ts
